+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [10.1.0](https://pypi.org/project/directory-cms-client/10.1.0/) (2019-06-18)
+[Full Changelog](https://github.com/uktrade/directory-cms-client/pull/46/files)
+
+**Implemented enhancements:**
+
+- Added `limit` and `offset` optional params to `list_by_page_type` to allow for pagination
+
+
 ## [10.0.1](https://pypi.org/project/directory-cms-client/10.0.1/) (2019-05-30)
 [Full Changelog](https://github.com/uktrade/directory-cms-client/pull/44/files)